+++ sdk/resourcemanager/compute/armcompute/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 6.5.0 (2025-05-06)
+### Features Added
+
+- New field `Tags` in struct `VirtualMachineNetworkInterfaceConfiguration`
+- New field `Tags` in struct `VirtualMachinePublicIPAddressConfiguration`
+- New field `Tags` in struct `VirtualMachineScaleSetNetworkConfiguration`
+- New field `Tags` in struct `VirtualMachineScaleSetPublicIPAddressConfiguration`
+- New field `Tags` in struct `VirtualMachineScaleSetUpdateNetworkConfiguration`
+- New field `Tags` in struct `VirtualMachineScaleSetUpdatePublicIPAddressConfiguration`
+- New field `PatchNameMasksToExclude`, `PatchNameMasksToInclude` in struct `WindowsParameters`
+
+
 ## 6.4.0 (2025-03-28)
 ### Features Added